category/views.py
from rest_framework import generics
from .models import Category, Brand
from .serializers import CategorySerializer, BrandSerializer


class CategoryView(generics.ListCreateAPIView):
    queryset = Category.objects.all()
    serializer_class = CategorySerializer


class BrandView(generics.ListCreateAPIView):
    queryset = Brand.objects.all()
    serializer_class = BrandSerializer
